Fix benchmark analyzer errors (MA0048, CA1001)
Move BenchmarkRecord to its own file and make ExtractorBenchmarks
sealed + IDisposable to satisfy code analysis rules.
