+++ benchmarks/Wolfgang.Etl.DbClient.Benchmarks/BenchmarkRecord.cs
@@ -1,0 +1,9 @@
+namespace Wolfgang.Etl.DbClient.Benchmarks;
+
+public class BenchmarkRecord
+{
+    public int Id { get; set; }
+    public string FirstName { get; set; } = string.Empty;
+    public string LastName { get; set; } = string.Empty;
+    public int Age { get; set; }
+}
